%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cdfGetGroups.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cdfGetGroups}
\alias{cdfGetGroups}


\title{Gets a subset of groups in a CDF structure}

\description{
 Gets a subset of groups in a CDF structure.

 This \code{\link[base]{function}} is designed to be used with \code{\link{applyCdfGroups}}().
}

\usage{
cdfGetGroups(groups, which, ...)
}

\arguments{
 \item{groups}{A \code{\link[base]{list}} of groups.}
 \item{which}{An \code{\link[base]{integer}} or \code{\link[base]{character}} \code{\link[base]{vector}} of groups be returned.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure of groups.
}

\seealso{
 \code{\link{applyCdfGroups}}().
}

\author{Henrik Bengtsson}



\keyword{programming}
\keyword{internal}
