%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  cdfSetDimension.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{cdfSetDimension}
\alias{cdfSetDimension}


\title{Sets the dimension of an object}

\description{
 Sets the dimension of an object.

 This \code{\link[base]{function}} is designed to be used with \code{\link{applyCdfGroupFields}}().
}

\usage{
cdfSetDimension(field, dim, ...)
}

\arguments{
 \item{groups}{A \code{\link[base]{list}} of groups.}
 \item{which}{An \code{\link[base]{integer}} or \code{\link[base]{character}} \code{\link[base]{vector}} of groups be returned.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure of groups.
}

\seealso{
 \code{\link{applyCdfGroupFields}}().
}

\author{Henrik Bengtsson}



\keyword{programming}
\keyword{internal}
