\name{readClfHeader}
\alias{readClfHeader}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Read the header of a CLF file.}
\description{
  Reads the header of a CLF file. The exact information stored in this
  file can be viewed in the readClfEnv documentation which reads the
  header in addition to the body.
}
\usage{
readClfHeader(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{\code{file} a CLF file}
}
\value{
  A list of header elements. 
}
\keyword{IO}
